# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PatentFlow BE API",
    "description" : "PatentFlow FE 우선 연동 API 문서",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "paths" : {
...
    "/api/v1/patents/{patentId}" : {
      "get" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "getPatentDetail",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentDetailResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "updatePatent",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatentUpsertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentUpsertResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patents" : {
      "get" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "getPatents",
        "parameters" : [
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 1
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 20
            }
          },
          {
            "name" : "keyword",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "departmentId",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "reviewWorkflowStatus",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "enum" : [
                "NOT_IN_REVIEW",
                "REVIEW_QUARTER_STARTED",
                "MAIL_READY",
                "WAITING_BUSINESS_RESPONSE",
                "BUSINESS_RESPONSE_RECEIVED"
              ]
            }
          },
          {
            "name" : "sort",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "quarter",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "country",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "dateFrom",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "format" : "date"
            }
          },
          {
            "name" : "dateTo",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "format" : "date"
            }
          },
          {
            "name" : "businessArea",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "technologyArea",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "productName",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "inReview",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "boolean"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponsePatentListItemResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "createPatent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatentUpsertRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentUpsertResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patents/{patentId}/request-ai-report" : {
      "post" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "requestAiReport",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAiReportJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patents/{patentId}/pdf" : {
      "get" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "downloadPatentPdf",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "uploadPatentPdf",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [
                  "file"
                ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentPdfMetaResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "deletePatentPdf",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentPdfMetaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patents/{patentId}/final-decision" : {
      "post" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "recordFinalDecision",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FinalDecisionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseFinalDecisionResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "patchFinalDecision",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatchFinalDecisionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseFinalDecisionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patents/{patentId}/co-applicant-consent" : {
      "post" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "recordCoApplicantConsent",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CoApplicantConsentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patents/context-suggestions" : {
      "post" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "suggestContext",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PatentContextSuggestionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentContextSuggestionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patents/{patentId}/department" : {
      "patch" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "assignDepartment",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssignDepartmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patents/{patentId}/ai-report" : {
      "patch" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "editAiReport",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AiReportEditRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAiEvaluationReportResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patents/bulk/department" : {
      "patch" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "bulkAssignDepartment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulkAssignDepartmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBulkAssignDepartmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/patents/{patentId}/pdf/meta" : {
      "get" : {
        "tags" : [
          "patent-controller"
        ],
        "operationId" : "getPatentPdfMeta",
        "parameters" : [
          {
            "name" : "patentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePatentPdfMetaResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponsePatentUpsertResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PatentUpsertResponse"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ApiResponsePatentDetailResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PatentDetailResponse"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageResponsePatentListItemResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PatentListItemResponse"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageInfo"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}