# src/main/java/com/syuuk/patentflow/patent/controller/PatentController.java
package com.syuuk.patentflow.patent.controller;

import com.syuuk.patentflow.auth.dto.UserPrincipalResponse;
import com.syuuk.patentflow.common.response.ApiResponse;
import com.syuuk.patentflow.common.response.PageResponse;
import com.syuuk.patentflow.patent.dto.AssignDepartmentRequest;
import com.syuuk.patentflow.patent.dto.BulkAssignDepartmentRequest;
import com.syuuk.patentflow.patent.dto.AiEvaluationReportResponse;
import com.syuuk.patentflow.patent.dto.AiReportEditRequest;
import com.syuuk.patentflow.patent.dto.AiReportJobResponse;
import com.syuuk.patentflow.patent.dto.CoApplicantConsentRequest;
import com.syuuk.patentflow.patent.dto.FinalDecisionRequest;
import com.syuuk.patentflow.patent.dto.FinalDecisionResponse;
import com.syuuk.patentflow.patent.dto.PatchFinalDecisionRequest;
import com.syuuk.patentflow.patent.dto.PatentBibliographicInfoResponse;
import com.syuuk.patentflow.patent.dto.PatentContextSuggestionRequest;
import com.syuuk.patentflow.patent.dto.PatentContextSuggestionResponse;
import com.syuuk.patentflow.patent.dto.PatentDetailResponse;
import com.syuuk.patentflow.patent.dto.PatentFilterOptionsResponse;
import com.syuuk.patentflow.patent.dto.PatentHistoryResponse;
import com.syuuk.patentflow.patent.dto.PatentFeeScheduleResponse;
import com.syuuk.patentflow.patent.dto.PatentListFilter;
import com.syuuk.patentflow.patent.dto.PatentListItemResponse;
import com.syuuk.patentflow.patent.dto.PatentPdfMetaResponse;

import com.syuuk.patentflow.patent.dto.PatentUpsertRequest;
import com.syuuk.patentflow.patent.dto.PatentUpsertResponse;
import com.syuuk.patentflow.patent.dto.ReviewWorkflowStatus;
import com.syuuk.patentflow.patent.service.AiReportEditService;
import com.syuuk.patentflow.patent.service.AiReportJobService;
import com.syuuk.patentflow.patent.service.AnnualFeeScheduleManagementService;
import com.syuuk.patentflow.patent.service.PatentPdfService;
import com.syuuk.patentflow.patent.service.PatentReviewService;
import jakarta.validation.Valid;
import java.io.IOException;
import java.time.LocalDate;
import java.util.List;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/patents")
public class PatentController {

    private final PatentReviewService patentReviewService;
    private final AiReportJobService aiReportJobService;
    private final AiReportEditService aiReportEditService;
    private final AnnualFeeScheduleManagementService annualFeeScheduleManagementService;
    private final PatentPdfService patentPdfService;

    public PatentController(
            PatentReviewService patentReviewService,
            AiReportJobService aiReportJobService,
            AiReportEditService aiReportEditService,
            AnnualFeeScheduleManagementService annualFeeScheduleManagementService,
            PatentPdfService patentPdfService
    ) {
        this.patentReviewService = patentReviewService;
        this.aiReportJobService = aiReportJobService;
        this.aiReportEditService = aiReportEditService;
        this.annualFeeScheduleManagementService = annualFeeScheduleManagementService;
        this.patentPdfService = patentPdfService;
    }

    /**
     * @relatedFR FR-LEGAL-01, FR-LEGAL-02
     * @relatedUI UI-COM-02, UI-LEGAL-02
     * @description 특허 목록 검색/필터링/정렬/페이징 API.
     */
    @GetMapping
    public PageResponse<PatentListItemResponse> getPatents(
            @RequestParam(defaultValue = "1") int page,
            @RequestParam(defaultValue = "20") int size,
            @RequestParam(required = false) String keyword,
            @RequestParam(required = false) String departmentId,
            @RequestParam(required = false) ReviewWorkflowStatus reviewWorkflowStatus,
            @RequestParam(required = false) String sort,
            @RequestParam(required = false) String quarter,
            @RequestParam(required = false) String country,
            @RequestParam(required = false) LocalDate dateFrom,
            @RequestParam(required = false) LocalDate dateTo,
            @RequestParam(required = false) String businessArea,
            @RequestParam(required = false) String technologyArea,
            @RequestParam(required = false) String productName,
            @RequestParam(required = false) Boolean inReview) {
        // CONTRACT-09/DASH-08: 분기/국가/날짜/영역/검토여부까지 서버 필터로 위임받아 페이징 결과를 반환한다.
        return patentReviewService.getPatents(page, size, sort, new PatentListFilter(
                keyword, departmentId, reviewWorkflowStatus, quarter, country, dateFrom, dateTo,
                businessArea, technologyArea, productName, inReview));
    }

    /**
     * @relatedFR FR-LEGAL-01, FR-LEGAL-02
     * @relatedUI UI-LEGAL-01, UI-COM-02
     * @description CONTRACT-09/DASH-08: 검토 대상 목록 화면의 필터 드롭다운(국가·사업·기술·제품) 옵션을
     *     전체 특허 기준 distinct 값으로 제공한다(서버 필터로 목록이 좁혀져도 옵션 불변).
     */
    @GetMapping("/filter-options")
    public ApiResponse<PatentFilterOptionsResponse> getFilterOptions() {
        return ApiResponse.ok(patentReviewService.getFilterOptions());
    }

    /**
     * @relatedFR FR-LEGAL-22, FR-LEGAL-24
     * @relatedUI UI-LEGAL-01, UI-COM-02, UI-BUS-01
     * @description 분기/날짜 범위/국가 기준 검토 대상 특허를 조회한다.
     */
    @GetMapping("/review-targets")
    public ApiResponse<List<PatentListItemResponse>> getReviewTargets(
            @RequestParam(required = false) String quarter,
            @RequestParam(required = false) String country,
            @RequestParam(required = false) LocalDate dateFrom,
            @RequestParam(required = false) LocalDate dateTo,
            @RequestParam(required = false) ReviewWorkflowStatus reviewWorkflowStatus) {
        return ApiResponse.ok(patentReviewService.getReviewTargets(quarter, country, dateFrom, dateTo, reviewWorkflowStatus));
    }

    /**
     * @relatedFR FR-LEGAL-03, FR-LEGAL-04
     * @relatedUI UI-LEGAL-02, UI-LEGAL-03
     * @description 특허 기본 정보와 회사 컨텍스트 등록 API.
     */
    @PostMapping
    public ApiResponse<PatentUpsertResponse> createPatent(@Valid @RequestBody PatentUpsertRequest request) {
        return ApiResponse.ok(patentReviewService.createPatent(request));
    }

    /**
     * @relatedFR FR-LEGAL-03
     * @relatedUI UI-LEGAL-02, UI-LEGAL-03
     * @description 관리번호/출원번호/등록번호 기반 특허 외부 검색 API.
     */
    @GetMapping("/external-lookup")
    public ApiResponse<PatentBibliographicInfoResponse> lookupBibliographicInfo(
            @RequestParam(required = false) String managementNumber,
            @RequestParam(required = false) String applicationNumber,
            @RequestParam(required = false) String registrationNumber,
            @RequestParam(required = false) String sourcePriority
    ) {
        return ApiResponse.ok(patentReviewService.lookupBibliographicInfo(
                managementNumber,
                applicationNumber,
                registrationNumber,
                sourcePriority));
    }

    /**
     * @relatedFR FR-LEGAL-03, FR-LEGAL-04
     * @relatedUI UI-LEGAL-02, UI-LEGAL-03
     * @description 특허명/관련제품 기반 회사 컨텍스트 추천 API.
     */
    @PostMapping("/context-suggestions")
    public ApiResponse<PatentContextSuggestionResponse> suggestContext(
            @RequestBody PatentContextSuggestionRequest request
    ) {
        return ApiResponse.ok(patentReviewService.suggestContext(request));
    }

    /**
     * @relatedFR FR-LEGAL-05, FR-LEGAL-06, FR-LEGAL-07, FR-LEGAL-08, FR-LEGAL-09, FR-LEGAL-10
     * @relatedUI UI-LEGAL-04
     * @description 특허 상세, AI 평가 레포트, 최종 판단 분리 조회 API.
     */
    @GetMapping("/{patentId}")
    public ApiResponse<PatentDetailResponse> getPatentDetail(@PathVariable String patentId) {
        return ApiResponse.ok(patentReviewService.getPatentDetail(patentId));
    }

    /**
     * @relatedFR FR-LEGAL-24
     * @relatedUI UI-LEGAL-04
     * @description FEE-06: 특허 상세 연차료 일정 — 국가 규칙 기반 도래일·검토 시작(고지 발송) 예정일·수신처 조회 API.
     */
    @GetMapping("/{patentId}/fee-schedule")
    public ApiResponse<PatentFeeScheduleResponse> getPatentFeeSchedule(@PathVariable String patentId) {
        return ApiResponse.ok(annualFeeScheduleManagementService.getPatentFeeSchedule(patentId));
    }

    /**
     * @relatedFR FR-LEGAL-05
     * @description F6: 특허 패밀리(같은 관리번호 계열의 국가별 출원) 조회.
     */
    @GetMapping("/{patentId}/family")
    public ApiResponse<List<PatentListItemResponse>> getPatentFamily(@PathVariable String patentId) {
        return ApiResponse.ok(patentReviewService.getPatentFamily(patentId));
    }

    /**
     * @relatedFR FR-LEGAL-13
     * @relatedUI UI-LEGAL-03
     * @description MAIL-13: 법무팀 특허 PDF 직접 업로드 — TW·UAE 등 KIPRIS로 공개전문을
     *     가져올 수 없는 국가의 특허에 등록/수정 화면에서 PDF를 첨부한다(기존 첨부는 교체).
     */
    @PostMapping(value = "/{patentId}/pdf", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
    public ApiResponse<PatentPdfMetaResponse> uploadPatentPdf(
            @PathVariable String patentId,
            @RequestParam("file") MultipartFile file,
            Authentication authentication
    ) throws IOException {
        return ApiResponse.ok(patentPdfService.upload(
                patentId, file.getOriginalFilename(), file.getBytes(), currentActor(authentication)));
    }

    /**
     * @relatedFR FR-LEGAL-13
     * @description MAIL-13: 업로드된 특허 PDF 다운로드(관리자).
     */
    @GetMapping("/{patentId}/pdf")
    public ResponseEntity<byte[]> downloadPatentPdf(@PathVariable String patentId) {
        return pdfDownloadResponse(patentPdfService.downloadUploaded(patentId));
    }

    /**
     * @relatedFR FR-LEGAL-13
     * @description MAIL-13: 특허 PDF 첨부 상태 조회 — 등록/수정 화면 업로드 위젯이 사용한다.
     */
    @GetMapping("/{patentId}/pdf/meta")
    public ApiResponse<PatentPdfMetaResponse> getPatentPdfMeta(@PathVariable String patentId) {
        return ApiResponse.ok(patentPdfService.meta(patentId));
    }

    /**
     * @relatedFR FR-LEGAL-13
     * @description MAIL-13: 업로드된 특허 PDF 삭제(교체·정리용).
     */
    @DeleteMapping("/{patentId}/pdf")
    public ApiResponse<PatentPdfMetaResponse> deletePatentPdf(@PathVariable String patentId) {
        patentPdfService.deleteUploaded(patentId);
        return ApiResponse.ok(PatentPdfMetaResponse.none(patentId));
    }

    public static ResponseEntity<byte[]> pdfDownloadResponse(PatentPdfService.PdfDownload download) {
        String encodedName = java.net.URLEncoder.encode(download.docName(), java.nio.charset.StandardCharsets.UTF_8)
                .replace("+", "%20");
        return ResponseEntity.ok()
                .header("Content-Disposition", "attachment; filename*=UTF-8''" + encodedName)
                .contentType(MediaType.APPLICATION_PDF)
                .body(download.content());
    }

    /**
     * @relatedFR FR-LEGAL-03, FR-LEGAL-04
     * @relatedUI UI-LEGAL-02, UI-LEGAL-03
     * @description 특허 기본 정보와 회사 컨텍스트 수정 API.
     */
    @PutMapping("/{patentId}")
    public ApiResponse<PatentUpsertResponse> updatePatent(
            @PathVariable String patentId,
            @Valid @RequestBody PatentUpsertRequest request
    ) {
        return ApiResponse.ok(patentReviewService.updatePatent(patentId, request));
    }

    /**
     * @relatedFR FR-LEGAL-11
     * @relatedUI UI-LEGAL-04, UI-BUS-05
     * @description 평가/판단 이력 조회 API.
     */
    @GetMapping("/{patentId}/history")
    public ApiResponse<List<PatentHistoryResponse>> getPatentHistory(@PathVariable String patentId) {
        return ApiResponse.ok(patentReviewService.getPatentHistory(patentId));
    }

    /**
     * @relatedFR FR-LEGAL-09, FR-LEGAL-10
     * @relatedUI UI-LEGAL-04
     * @description 특허 최종 판단을 기록하는 API.
     */
    @PostMapping("/{patentId}/final-decision")
    public ApiResponse<FinalDecisionResponse> recordFinalDecision(
            @PathVariable String patentId,
            @Valid @RequestBody FinalDecisionRequest request,
            Authentication authentication
    ) {
        return ApiResponse.ok(patentReviewService.recordFinalDecision(patentId, request, currentActor(authentication)));
    }

    /**
     * @relatedFR FR-LEGAL-09, FR-LEGAL-10
     * @relatedUI UI-LEGAL-04
     * @description 공동출원 특허의 공동출원인 합의를 기록하는 API. 연차료 유지/포기 최종 판단의 전제 게이트.
     */
    @PostMapping("/{patentId}/co-applicant-consent")
    public ApiResponse<PatentDetailResponse> recordCoApplicantConsent(
            @PathVariable String patentId,
            @Valid @RequestBody CoApplicantConsentRequest request,
            Authentication authentication
    ) {
        return ApiResponse.ok(patentReviewService.recordCoApplicantConsent(patentId, request, currentActor(authentication)));
    }

    /**
     * @relatedFR FR-LEGAL-20
     * @relatedUI UI-LEGAL-04
     * @description 특허 최종 판단을 수정하거나 취소하는 API.
     */
    @PatchMapping("/{patentId}/final-decision")
    public ApiResponse<FinalDecisionResponse> patchFinalDecision(
            @PathVariable String patentId,
            @RequestBody PatchFinalDecisionRequest request,
            Authentication authentication
    ) {
        return ApiResponse.ok(patentReviewService.patchFinalDecision(patentId, request, currentActor(authentication)));
    }

    /**
     * @description 이번 분기 납부 대상 특허에 담당 사업부를 배정한다.
     */
    @PatchMapping("/{patentId}/department")
    public ApiResponse<PatentDetailResponse> assignDepartment(
            @PathVariable String patentId,
            @Valid @RequestBody AssignDepartmentRequest request
    ) {
        return ApiResponse.ok(patentReviewService.assignDepartment(patentId, request.departmentId()));
    }

    /**
     * @relatedFR FR-LEGAL-02
     * @description F5: 특허 다건 부서 일괄 배정 — 건별 실패는 건너뛰고 성공/실패 ID 목록을 반환한다.
     */
    @PatchMapping("/bulk/department")
    public ApiResponse<BulkAssignDepartmentResponse> bulkAssignDepartment(
            @Valid @RequestBody BulkAssignDepartmentRequest request
    ) {
        List<String> assigned = new java.util.ArrayList<>();
        List<String> failed = new java.util.ArrayList<>();
        for (String patentId : request.patentIds()) {
            try {
                patentReviewService.assignDepartment(patentId, request.departmentId());
                assigned.add(patentId);
            } catch (Exception exception) {
                failed.add(patentId);
            }
        }
        return ApiResponse.ok(new BulkAssignDepartmentResponse(assigned, failed));
    }

    public record BulkAssignDepartmentResponse(List<String> assignedPatentIds, List<String> failedPatentIds) {}

    /**
     * @description AI 평가 레포트 생성을 비동기 잡으로 요청한다.
     */
    @PostMapping("/{patentId}/request-ai-report")
    public ResponseEntity<ApiResponse<AiReportJobResponse>> requestAiReport(@PathVariable String patentId) {
        return ResponseEntity.status(HttpStatus.ACCEPTED).body(ApiResponse.ok(aiReportJobService.requestAiReport(patentId)));
    }

    /**
     * @relatedFR FR-LEGAL-09
     * @relatedUI UI-LEGAL-04
     * @description 법무팀이 AI 평가 레포트(권고/사유/축별 점수/판단근거/markdown)를 편집한다.
     *     AI 원본은 불변 보존되고 편집은 오버라이드로 분리 저장된다(부분 PATCH 누적).
     */
    @PatchMapping("/{patentId}/ai-report")
    public ApiResponse<AiEvaluationReportResponse> editAiReport(
            @PathVariable String patentId,
            @Valid @RequestBody AiReportEditRequest request,
            Authentication authentication
    ) {
        return ApiResponse.ok(aiReportEditService.editAiReport(patentId, request, currentActor(authentication)));
    }

    /**
     * @relatedFR FR-LEGAL-09
     * @description 법무 편집을 모두 폐기하고 AI 원본 레포트로 되돌린다.
     */
    @DeleteMapping("/{patentId}/ai-report/edits")
    public ApiResponse<AiEvaluationReportResponse> revertAiReportEdits(
            @PathVariable String patentId,
            Authentication authentication
    ) {
        return ApiResponse.ok(aiReportEditService.revertAiReport(patentId, currentActor(authentication)));
    }

    /**
     * @relatedFR FR-LEGAL-09
     * @description 'AI 원본 보기' — 법무 편집을 반영하지 않은 순수 AI 레포트를 조회한다.
     */
    @GetMapping("/{patentId}/ai-report/original")
    public ApiResponse<AiEvaluationReportResponse> getOriginalAiReport(@PathVariable String patentId) {
        return ApiResponse.ok(patentReviewService.getOriginalAiReport(patentId));
    }

    /**
     * @description 최신 AI 평가 레포트 생성 잡 상태를 조회한다.
     */
    @GetMapping("/{patentId}/ai-report/status")
    public ApiResponse<AiReportJobResponse> getAiReportStatus(@PathVariable String patentId) {
        return ApiResponse.ok(aiReportJobService.latestStatus(patentId));
    }

    // REVIEW-07: 최종 판단 행위자(인증 주체)를 추출한다. 미인증이면 '관리자'로 폴백.
    private String currentActor(Authentication authentication) {
        if (authentication == null || !authentication.isAuthenticated()) {
            return "관리자";
        }
        Object principal = authentication.getPrincipal();
        if (principal instanceof UserPrincipalResponse userPrincipal) {
            if (userPrincipal.username() != null && !userPrincipal.username().isBlank()) {
                return userPrincipal.username().trim();
            }
            if (userPrincipal.email() != null && !userPrincipal.email().isBlank()) {
                return userPrincipal.email().trim();
            }
        }
        String name = authentication.getName();
        return name == null || name.isBlank() ? "관리자" : name.trim();
    }

}